feat(tests): add integration tests for command execution and version retrieval
